+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Polyverse Boost Cloud Service
 ### Enhancements
 - Draft Blueprint service returns a prioritized list of the files to analyze
 - Speed up Draft Blueprint service by using 3.5 Turbo model
+- Added Sara identity to all system prompts, using prompt injection at runtime - overridable by client using 'system_identity' data parameter
 
 ### Bug Fixes
 - N/A